+++ features/step_definitions/docker_command_steps.rb
@@ -14,6 +14,8 @@ Given(/^the docker command finds a running shell "(.*?)"$/) do |shell_name|
   expect_any_instance_of(Geordi::Docker).to receive(:mock_parse).at_least(:once).and_wrap_original do |original, *args|
     if args[0] =~ /docker-compose ps/
       "other_shell\n#{shell_name} foo"
+    elsif args[1] =~ /docker-compose ps/
+      shell_name
     else
       original.call(*args)
     end